Set up 2FA route for qr generation
If users haven't got the two factor authentication enabled the final
jwt tokens (access and refresh) are created and he gets the access to the profile.
If he has 2fa enabled he will be redirected to the 2fa route in the front-end to initiate
the 2fa flow (qr code generation and code verification).